#1838 moved api.js and auth.js to utilities dir
